stores old ReviewAllView view in 'ReviewAllView old.tsx' file and has a simple ReviewAllView that contains the review_all_view_intro text
