+++ vite-react/package.json
@@ -16,7 +16,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.474.0",
-    "pdfjs-dist": "^5.1.91",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-native-dotenv": "^3.4.11",
