+++ vite-react/src/views/ReviewAllView.tsx
@@ -1,9 +1,7 @@
-// src/views/ReviewSectionView.tsx
 import { useEffect } from "react";
 import { useAppContext } from "@/context/AppContext";
 import { useNavigate } from "react-router-dom";
 import { Card } from "@/components/ui/card";
-import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 
 export function ReviewAllView() {
   const navigate = useNavigate();
@@ -17,54 +15,14 @@ export function ReviewAllView() {
 
   if (!letterLabData) return null;
 
-  const bullets = {
-    enactive_mastery: Object.values(letterLabData.BSETB_enactive_mastery).map(
-      (b) => b.text
-    ),
-    vicarious_experience: Object.values(
-      letterLabData.BSETB_vicarious_experience
-    ).map((b) => b.text),
-    verbal_persuasion: Object.values(letterLabData.BSETB_verbal_persuasion).map(
-      (b) => b.text
-    ),
-  };
-
   return (
-    <Card className="w-full max-w-6xl p-6 bg-white shadow-lg">
-      <Tabs defaultValue="cover">
-        <TabsList className="flex border-b mb-4">
-          <TabsTrigger value="cover">Cover Letter</TabsTrigger>
-          <TabsTrigger value="bullets">Bullet Points</TabsTrigger>
-        </TabsList>
-
-        <TabsContent value="cover">
-          <h2 className="text-xl font-semibold mb-2">Cover Letter</h2>
-          <p className="whitespace-pre-line">
-            {letterLabData.initial_cover_letter}
-          </p>
-        </TabsContent>
-
-        <TabsContent value="bullets">
-          <h2 className="text-xl font-semibold mb-2">
-            Bullet Points by Bandura Category
-          </h2>
-
-          <div className="mt-4">
-            {Object.entries(bullets).map(([key, points]) => (
-              <div key={key} className="mb-6">
-                <h3 className="text-lg font-bold capitalize underline">
-                  {key.replace("_", " ")}
-                </h3>
-                <ul className="list-disc ml-6 mt-2 space-y-2">
-                  {points.map((bp, idx) => (
-                    <li key={idx}>{bp}</li>
-                  ))}
-                </ul>
-              </div>
-            ))}
-          </div>
-        </TabsContent>
-      </Tabs>
+    <Card className="w-full max-w-4xl p-6 bg-white shadow-lg">
+      <h2 className="text-2xl font-bold mb-4">
+        Your Personalized Cover Letter Review
+      </h2>
+      <p className="text-gray-700 whitespace-pre-line">
+        {letterLabData.review_all_view_intro}
+      </p>
     </Card>
   );
 }
